# Fixed #26891 -- Fixed lookup registration for ForeignObject