https://code.djangoproject.com/ticket/26891

Update - just recapping the current state of the PR:

### Implementation:
The scope has expanded a bit. In fixing the original lookup registration issue, it was necessary to add `get_lookups()` to the registration API.

- `get_lookups()` simply returns the merged set of `class_lookups` in a field's class hierarchy. 
- For performance reasons, the lookups are cached on the field as `cached_lookups`. 
- Similar to registering lookups, `cached_lookups` are checked on the `cls.__dict__`. This prevents DateTimeField from accidentally using `Field.cached_lookups`, which wouldn't include the date transforms.
- `register_lookup()` will bust the cached lookups on a field and any of its subclasses. e.g., `ForeignObject.register_lookup()` would bust the cache of ForeignObject, ForeignKey, OneToOneField, etc...
- `ForeignObject.get_lookups()` only caches `class_lookups` up to it's position in the class hierarchy, excluding the lookups registered to Field.

### Performance tests:
Used timeit to time `field.get_lookup('in')` for different fields, comparing the current master to the PR. 100,000 iterations per run, 5 runs per field.

**master branch:**
```
CharField       average: 0.2338s [0.2377979755401611, 0.2334151268005371, 0.2305140495300293, 0.2334628105163574, 0.2340850830078125]
DateTimeField   average: 0.4872s [0.4839980602264404, 0.4864580631256103, 0.4956321716308594, 0.4867660999298095, 0.4834890365600586]
ForeignKey      average: 0.0266s [0.0261549949645996, 0.0266189575195312, 0.0267159938812255, 0.0269579887390136, 0.0267570018768310]
```

- Getting lookups on ForeignKey is fast because the lookups are hardcoded (which causes lookup registration to have no effect).
- DateTimeField is slower than CharField since it needs to test multiple `class_lookups` dicts.

**pr:**
```
CharField       average: 0.2818s [0.2849998474121094, 0.2799849510192871, 0.2836630344390869, 0.2778408527374267, 0.2827849388122558]
DateTimeField   average: 0.2833s [0.2864518165588379, 0.2799329757690429, 0.2777631282806396, 0.2892100811004638, 0.2832741737365722]
ForeignKey      average: 0.2851s [0.2807130813598633, 0.2836570739746094, 0.2891471385955810, 0.2907619476318359, 0.2814109325408935]
```