@timgraham 2016-07-13
Thanks for the patch. All non-trivial pull requests also require a Trac ticket.