@rpkilby 2016-07-13
will do - should I just edit the PR once the Trac ticket is open?