@jarshwah 2016-07-15
Looks mostly good to me. I'm not very familiar with `ForeignObject` though so I might have missed something obvious. Was there a particular lookup you were trying to create that led you to fixing this? The test case you've got (`Exactly`) shows the feature working, but not really why the changed needed to be made.