@rpkilby 2016-08-01
The blank lines have been removed.

My thoughts on updating the docs would be if `get_lookups()` was to be considered as public API or not.