@timgraham 2016-08-10
You can squash commits when updating also. Thanks!