@rpkilby 2016-08-11
Copying here since the diff is outdated:

----

> Could it be useful to use `@functools.lru_cache()` or `@cached_property` rather than implementing your own caching?

The caching here is based on the same idea as `class_lookups`, just a bit more complicated because of the cache busting. I didn't try `@cached_property` (no way to bust the cached value, and it doesn't work w/ classmethods, right?), but `lru_cache` worked just fine. That said, it is about 50% slower.
  - current implementation at 100k runs takes ~.28s on average.
  - `lru_cache` at 100k runs takes ~.42s on average.

50% is a significant percentage, but we're talking about lookup resolution taking .0000042 seconds vs .0000028 seconds. `¯\_(ツ)_/¯`

Either way, I've squashed the existing commits into one, and added the lru changes separately. Thoughts on the performance vs code clarity?