@timgraham 2016-08-11
I don't have a sense as to whether or not those performance differences would have any real-world impact.

If we keep the changes, I think you can drop `maxsize=None` since the methods don't have any arguments, the size shouldn't be larger than 1, right?