@rpkilby 2016-08-11
> I don't have a sense as to whether or not those performance differences would have any real-world impact.

Yeah. With my laptop, it take 700,000 lookups to use one additional second of CPU time. I feel like this is probably fine. Went ahead and squashed the commits.

> If we keep the changes, I think you can drop maxsize=None since the methods don't have any arguments, the size shouldn't be larger than 1, right?

No - it should be caching per field class. eg, the datetime lookups vs the standard lookups.